+++ src/fuzzy_asteroids/util.py
@@ -12,7 +12,7 @@ python -m arcade.examples.asteroids
 import random
 from typing import Any, Dict, List, Tuple
 
-from .sprites import AsteroidSprite
+from .sprites import AsteroidSprite, ShipSprite
 from .settings import SCREEN_WIDTH, SCREEN_HEIGHT
 
 
@@ -26,6 +26,7 @@ class Score:
     *  timestep_update()
     *  final_update()
     """
+
     def __init__(self):
         """
         Default constructor requires no arguments as all values are initialized to default values, which
@@ -70,7 +71,7 @@ class Score:
 
     @property
     def fraction_distance_travelled(self):
-        return 0.0 if not self.max_distance else self.distance_travelled/self.max_distance
+        return 0.0 if not self.max_distance else self.distance_travelled / self.max_distance
 
     def timestep_update(self, environment) -> None:
         """
@@ -99,6 +100,7 @@ class Map:
     """
     Class to be used to customize the game map
     """
+
     def __init__(self, width: int = SCREEN_WIDTH, height: int = SCREEN_HEIGHT):
         """
         Class for specifying the Map dimensions (visible + otherwise)
@@ -117,7 +119,7 @@ class Map:
 
     @property
     def center(self) -> Tuple[float, float]:
-        return self.width/2.0, self.height/2.0
+        return self.width / 2.0, self.height / 2.0
 
     @staticmethod
     def default_width() -> float:
@@ -134,12 +136,13 @@ class Map:
     @staticmethod
     def default_map_center() -> Tuple[float, float]:
         dims = Map.default_dimensions()
-        return dims[0]/2.0, dims[1]/2.0
+        return dims[0] / 2.0, dims[1] / 2.0
 
 
 class Scenario:
-    def __init__(self, name: str = "", num_asteroids: int = 0, asteroid_states: List[Dict[str, Any]] = None,
-                 ship_state: Dict[str, Any] = None, game_map: Map = None, seed: int = None):
+    def __init__(self, name: str = "Unnamed", num_asteroids: int = 0, asteroid_states: List[Dict[str, Any]] = None,
+                 ship_states: List[Dict[str, Any]] = None, game_map: Map = None, seed: int = None,
+                 time_limit: float = float("inf")):
         """
         Specify the starting state of the environment, including map dimensions and optional features
 
@@ -152,16 +155,24 @@ class Scenario:
         :param ship_state: Optional, Ship Starting state
         :param game_map: Game Map using ``Map`` object
         :param seed: Optional seeding value to pass to random.seed() which is called before asteroid creation
+        :param time_limit: Optional seeding value to pass to random.seed() which is called before asteroid creation
         """
         self.name = name
-        self.asteroid_states = list()
 
         # Store Map
         self.game_map = game_map if game_map else Map()
-        self.ship_state = ship_state if ship_state else {"position": self.game_map.center, "angle": 0.0}
 
-        # Store seed
+        # Store ship states if not None, otherwise, create one ship at center
+        self.ship_states = ship_states if ship_states else [{"position": self.game_map.center}]
+
+        # Set the time_limit to infinity if it is 0 or None
+        self.time_limit = time_limit
+
+        # Store random seed
         self.seed = seed
+
+        # Will be built later
+        self.asteroid_states = list()
 
         # Check for mismatch between explicitly defined number of asteroids and Tuple of states
         if num_asteroids and asteroid_states:
@@ -174,8 +185,8 @@ class Scenario:
         elif num_asteroids:
             self.asteroid_states = [dict() for _ in range(num_asteroids)]
         else:
-            raise(ValueError("User should define `num_asteroids` or `asteroid_states` to create "
-                             "valid custom starting states for the environment"))
+            raise (ValueError("User should define `num_asteroids` or `asteroid_states` to create "
+                              "valid custom starting states for the environment"))
 
     @property
     def num_starting_asteroids(self) -> float:
@@ -192,7 +203,7 @@ class Scenario:
     @staticmethod
     def count_asteroids(asteroid_size) -> float:
         # Counting based off of each asteroid making 3 children when destroyed
-        return sum([3 ** (size - 1) for size in range(1, asteroid_size+1)])
+        return sum([3 ** (size - 1) for size in range(1, asteroid_size + 1)])
 
     def asteroids(self, frequency: float) -> List[AsteroidSprite]:
         """
@@ -211,9 +222,20 @@ class Scenario:
             if asteroid_state:
                 asteroids.append(AsteroidSprite(frequency, **asteroid_state))
             else:
-                asteroids.append(AsteroidSprite(frequency,
-                                                position=(random.randrange(self.game_map.LEFT_LIMIT, self.game_map.RIGHT_LIMIT),
-                                                          random.randrange(self.game_map.BOTTOM_LIMIT, self.game_map.TOP_LIMIT)),
-                                                ))
+                asteroids.append(
+                    AsteroidSprite(frequency,
+                                   position=(
+                                       random.randrange(self.game_map.LEFT_LIMIT, self.game_map.RIGHT_LIMIT),
+                                       random.randrange(self.game_map.BOTTOM_LIMIT, self.game_map.TOP_LIMIT)),
+                                   ))
 
         return asteroids
+
+    def ships(self, frequency: float) -> List[ShipSprite]:
+        """
+        Create ship sprites
+        :param frequency: Operating frequency of the game
+        :return: List of ShipSprites
+        """
+        # Loop through and create ShipSprites based on starting state
+        return [ShipSprite(idx+1, frequency, **ship_state) for idx, ship_state in enumerate(self.ship_states)]